fix: don't attempt to bind -h if already bound
Binding a flag that already exists leads to a `panic`. While we did
check whether `--help` was bound, we did not check for the edge case
where `--help` is indeed unbound, but its shorthand `-h` was not.

This is now safely checked, so when the edge-case occurs, only the long
form will be bound.